refactored: /posts/[postId]
