+++ src/app/page.tsx
@@ -1,16 +1,10 @@
-import AboutMe from './components/AboutMe'
-import EmailSignUp from './components/EmailSignUp'
-import MyProfilePic from './components/MyProfilePic'
-import Posts from './components/Posts'
-import ProfileInfo from './components/ProfileInfo'
+import Link from "next/link";
+import ProfileInfo from "./components/ProfileInfo";
 
 export default function Home() {
   return (
-    <main> 
-      
-      <ProfileInfo/>
-      <Posts/>
-      {/* <EmailSignUp/> */}
+    <main>
+      <ProfileInfo />
     </main>
-  )
+  );
 }
