+++ src/lib/getFormattedDate.ts
@@ -1,4 +1,4 @@
 export default function getFormattedDate(dateString: string): string {
-    return new Intl.DateTimeFormat('en-US', { dateStyle: 'long' })
+    return new Intl.DateTimeFormat('en-US', { dateStyle: 'full' })
     .format(new Date(dateString))
 }
